+++ src/test/java/com/cinchapi/ccl/CompilerJavaCCLogicTest.java
@@ -19,6 +19,7 @@ import com.cinchapi.ccl.grammar.ConjunctionSymbol;
 import com.cinchapi.ccl.grammar.DirectionSymbol;
 import com.cinchapi.ccl.grammar.ExpressionSymbol;
 import com.cinchapi.ccl.grammar.FunctionKeySymbol;
+import com.cinchapi.ccl.grammar.FunctionTokenSymbol;
 import com.cinchapi.ccl.grammar.FunctionValueSymbol;
 import com.cinchapi.ccl.grammar.OperatorSymbol;
 import com.cinchapi.ccl.grammar.OrderComponentSymbol;
@@ -35,13 +36,14 @@ import com.cinchapi.ccl.syntax.AndTree;
 import com.cinchapi.ccl.syntax.ConditionTree;
 import com.cinchapi.ccl.syntax.ConjunctionTree;
 import com.cinchapi.ccl.syntax.ExpressionTree;
+import com.cinchapi.ccl.syntax.FunctionTree;
 import com.cinchapi.ccl.syntax.OrTree;
 import com.cinchapi.ccl.syntax.OrderTree;
 import com.cinchapi.ccl.syntax.PageTree;
 import com.cinchapi.ccl.syntax.CommandTree;
 import com.cinchapi.ccl.type.Operator;
 import com.cinchapi.ccl.type.function.IndexFunction;
-import com.cinchapi.ccl.type.function.KeyCclFunction;
+import com.cinchapi.ccl.type.function.KeyConditionFunction;
 import com.cinchapi.ccl.type.function.KeyRecordsFunction;
 import com.cinchapi.ccl.util.NaturalLanguage;
 import com.cinchapi.ccl.type.function.ImplicitKeyRecordFunction;
@@ -1319,14 +1321,28 @@ public class CompilerJavaCCLogicTest {
         ExpressionSymbol expression = (ExpressionSymbol) tree.root();
         Assert.assertEquals("age", expression.key().toString());
         Assert.assertEquals("><", expression.operator().toString());
-        KeyCclFunction function = (KeyCclFunction) expression.values().get(0)
-                .value();
-        Assert.assertTrue(function.source() instanceof ExpressionTree);
-        ExpressionTree t = (ExpressionTree) function.source();
-        ExpressionSymbol root = (ExpressionSymbol) t.root();
-        Assert.assertEquals("age", root.key().key());
-        Assert.assertEquals(">", root.operator().operator().symbol());
-        Assert.assertEquals(10, root.values().get(0).value());
+        Assert.assertTrue(
+                expression.values().get(0) instanceof FunctionValueSymbol);
+        Assert.assertEquals("avg",
+                ((KeyConditionFunction) expression.values().get(0).value())
+                        .operation());
+        Assert.assertEquals("age",
+                ((KeyConditionFunction) expression.values().get(0).value()).key());
+
+        Assert.assertTrue((((KeyConditionFunction) expression.values().get(0).value())
+                .source()) instanceof ExpressionTree);
+        Assert.assertEquals("age",
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
+                        .values().get(0).value()).source()).root()).key()
+                        .toString());
+        Assert.assertEquals(">",
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
+                        .values().get(0).value()).source()).root()).operator()
+                        .toString());
+        Assert.assertEquals("10",
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
+                        .values().get(0).value()).source()).root()).values()
+                        .get(0).toString());
 
         Assert.assertEquals("1000", expression.values().get(1).toString());
     }
@@ -1380,23 +1396,23 @@ public class CompilerJavaCCLogicTest {
         Assert.assertTrue(
                 expression.values().get(0) instanceof FunctionValueSymbol);
         Assert.assertEquals("avg",
-                ((KeyCclFunction) expression.values().get(0).value())
+                ((KeyConditionFunction) expression.values().get(0).value())
                         .operation());
         Assert.assertEquals("age",
-                ((KeyCclFunction) expression.values().get(0).value()).key());
+                ((KeyConditionFunction) expression.values().get(0).value()).key());
 
-        Assert.assertTrue((((KeyCclFunction) expression.values().get(0).value())
+        Assert.assertTrue((((KeyConditionFunction) expression.values().get(0).value())
                 .source()) instanceof ExpressionTree);
         Assert.assertEquals("age",
-                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyCclFunction) expression
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
                         .values().get(0).value()).source()).root()).key()
                                 .toString());
         Assert.assertEquals("<",
-                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyCclFunction) expression
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
                         .values().get(0).value()).source()).root()).operator()
                                 .toString());
         Assert.assertEquals("30",
-                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyCclFunction) expression
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
                         .values().get(0).value()).source()).root()).values()
                                 .get(0).toString());
     }
@@ -1421,23 +1437,23 @@ public class CompilerJavaCCLogicTest {
         Assert.assertTrue(
                 expression.values().get(0) instanceof FunctionValueSymbol);
         Assert.assertEquals("avg",
-                ((KeyCclFunction) expression.values().get(0).value())
+                ((KeyConditionFunction) expression.values().get(0).value())
                         .operation());
         Assert.assertEquals("age",
-                ((KeyCclFunction) expression.values().get(0).value()).key());
+                ((KeyConditionFunction) expression.values().get(0).value()).key());
 
-        Assert.assertTrue((((KeyCclFunction) expression.values().get(0).value())
+        Assert.assertTrue((((KeyConditionFunction) expression.values().get(0).value())
                 .source()) instanceof ExpressionTree);
         Assert.assertEquals("age",
-                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyCclFunction) expression
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
                         .values().get(0).value()).source()).root()).key()
                                 .toString());
         Assert.assertEquals("<",
-                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyCclFunction) expression
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
                         .values().get(0).value()).source()).root()).operator()
                                 .toString());
         Assert.assertEquals("30",
-                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyCclFunction) expression
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction) expression
                         .values().get(0).value()).source()).root()).values()
                                 .get(0).toString());
     }
@@ -2004,6 +2020,143 @@ public class CompilerJavaCCLogicTest {
         PageSymbol page = (PageSymbol) ((CommandTree) tree).pageTree().root();
         Assert.assertEquals(2, page.offset());
         Assert.assertEquals(1, page.limit());
+    }
+
+    @Test
+    public void testImplicitKeyRecordFunctionTokenize() {
+        String ccl = "age | avg";
+
+        // Build expected queue
+        List<Object> expectedTokens = Lists.newArrayList();
+
+        expectedTokens.add(new FunctionKeySymbol(
+                new ImplicitKeyRecordFunction("avg", "age")));
+
+        // Generate queue
+        Compiler compiler = Compiler.create(COMPILER_PARSE_VALUE_FUNCTION,
+                COMPILER_PARSE_OPERATOR_FUNCTION);
+        AbstractSyntaxTree ast = compiler.parse(ccl);
+        List<Symbol> tokens = compiler.tokenize(ast);
+
+        Assert.assertEquals(expectedTokens, tokens);
+    }
+
+    @Test
+    public void testImplicitKeyRecordFunctionAbstractSyntaxTree() {
+        String ccl = "age | avg";
+
+        // Generate tree
+        Compiler compiler = Compiler.create(COMPILER_PARSE_VALUE_FUNCTION,
+                COMPILER_PARSE_OPERATOR_FUNCTION);
+        AbstractSyntaxTree tree = compiler.parse(ccl);
+
+        // Root node
+        Assert.assertTrue(tree instanceof FunctionTree);
+
+        FunctionTokenSymbol symbol = (FunctionTokenSymbol) tree.root();
+        Assert.assertEquals("age", symbol.function().key());
+        Assert.assertEquals("avg", symbol.function().operation());
+    }
+
+    @Test
+    public void testIndexFunctionAbstractSyntaxTree() {
+        String ccl = "avg(age)";
+
+        // Generate tree
+        Compiler compiler = Compiler.create(COMPILER_PARSE_VALUE_FUNCTION,
+                COMPILER_PARSE_OPERATOR_FUNCTION);
+        AbstractSyntaxTree tree = compiler.parse(ccl);
+
+        // Root node
+        Assert.assertTrue(tree instanceof FunctionTree);
+
+        FunctionTokenSymbol symbol = (FunctionTokenSymbol) tree.root();
+        Assert.assertEquals(IndexFunction.class, symbol.function().getClass());
+        Assert.assertEquals("age", symbol.function().key());
+        Assert.assertEquals("avg", symbol.function().operation());
+    }
+
+    @Test
+    public void testKeyCclFunctionAbstractSyntaxTree() {
+        String ccl = "avg(age, age > 3)";
+
+        // Generate tree
+        Compiler compiler = Compiler.create(COMPILER_PARSE_VALUE_FUNCTION,
+                COMPILER_PARSE_OPERATOR_FUNCTION);
+        AbstractSyntaxTree tree = compiler.parse(ccl);
+
+        // Root node
+        Assert.assertTrue(tree instanceof FunctionTree);
+        FunctionTokenSymbol symbol = (FunctionTokenSymbol) tree.root();
+
+        Assert.assertEquals("avg",
+                ((KeyConditionFunction) symbol.function()).operation());
+        Assert.assertEquals("age",
+                ((KeyConditionFunction) symbol.function()).key());
+
+        Assert.assertTrue((((KeyConditionFunction) symbol.function())
+                .source()) instanceof ExpressionTree);
+        Assert.assertEquals("age",
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction)
+                        symbol.function()).source()).root()).key().toString());
+        Assert.assertEquals(">",
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction)
+                        symbol.function()).source()).root()).operator()
+                        .toString());
+        Assert.assertEquals("3",
+                ((ExpressionSymbol) ((AbstractSyntaxTree) ((KeyConditionFunction)
+                        symbol.function()).source()).root()).values()
+                        .get(0).toString());
+    }
+
+    @Test
+    public void testKeyRecordsFunctionAbstractSyntaxTree() {
+        String ccl = "avg(age, 1)";
+
+        // Generate tree
+        Compiler compiler = Compiler.create(COMPILER_PARSE_VALUE_FUNCTION,
+                COMPILER_PARSE_OPERATOR_FUNCTION);
+        AbstractSyntaxTree tree = compiler.parse(ccl);
+
+        // Root node
+        Assert.assertTrue(tree instanceof FunctionTree);
+        FunctionTokenSymbol symbol = (FunctionTokenSymbol) tree.root();
+
+        Assert.assertEquals("avg",
+                ((KeyRecordsFunction) symbol.function()).operation());
+        Assert.assertEquals("age",
+                ((KeyRecordsFunction) symbol.function()).key());
+
+        Assert.assertEquals(1, ((List<String>) ((KeyRecordsFunction)
+                symbol.function()).source()).size());
+        Assert.assertEquals("1",
+                ((List<String>) ((KeyRecordsFunction)
+                        symbol.function()).source()).get(0));
+    }
+    
+    @Test
+    public void testKeyMultiRecordsFunctionAbstractSyntaxTree() {
+        String ccl = "avg(age, 1,2,3,5,11)";
+
+        // Generate tree
+        Compiler compiler = Compiler.create(COMPILER_PARSE_VALUE_FUNCTION,
+                COMPILER_PARSE_OPERATOR_FUNCTION);
+        AbstractSyntaxTree tree = compiler.parse(ccl);
+
+        // Root node
+        Assert.assertTrue(tree instanceof FunctionTree);
+        FunctionTokenSymbol symbol = (FunctionTokenSymbol) tree.root();
+
+        Assert.assertEquals("avg",
+                ((KeyRecordsFunction) symbol.function()).operation());
+        Assert.assertEquals("age",
+                ((KeyRecordsFunction) symbol.function()).key());
+
+        Assert.assertEquals(5, ((List<String>) ((KeyRecordsFunction)
+                symbol.function()).source()).size());
+        Assert.assertEquals("1",
+                ((List<String>) ((KeyRecordsFunction)
+                        symbol.function()).source()).get(0));
     }
 
     @Test